+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `deploy_token` support for deploying ERC20 tokens from MPC / dev-managed wallets.
+- Allow for calling `invoke_contract` with a SmartContract object.
 
 ## [0.5.0] - 2024-09-11